MEDIUM CODE RUB: Give Every Model Its Own File
Verified: build 0 Warning(s) 0 Error(s), test 198/198, conformance 6 passed.

Eighteen wire records were declared inside the four brokers that use them. They are
models and they now live under Models -> Brokers -> {Entity}:

  Models/Brokers/Generators   ChatCompletionRequest ChatMessage ChatCompletionResponse ChatChoice
  Models/Brokers/Classifiers  the same four
  Models/Brokers/Verifiers    the same four
  Models/Brokers/Mcps         JsonRpcRequest ToolCallParams JsonRpcResponse
                              ToolCallResult ToolCallContent JsonRpcError
  Models/Clients/Agents       InferenceSettings

The three chat brokers each keep their own copy rather than sharing one. That is
deliberate: The Standard prefers duplication over a model shared across independent
flows, and swapping the Generator's protocol must not touch the Gate's. Same type
names, three namespaces, no coupling.

They were private nested records; at namespace level they are internal. The protocol
still cannot leak to consumers, which is what private was doing.

Two more that were not in the report:

InferenceSettings was nested in StandardAgent.

Harness.cs held ELEVEN types — eight stub brokers, a tool and two models — in one
file. Split into Conformance/Brokers, Conformance/Tools and Conformance/Models. 0.0
asks for a file per type and that file was the worst offender in the repo.

Final sweep: 0 nested types, 0 files declaring more than one top-level type.

Note worth a look: Conformance/Brokers/NotConfiguredMcpBroker.cs is now a near-duplicate
of the library's own Brokers/Mcps/NotConfiguredMcpBroker.cs from #81 — same name, same
behaviour, different namespace. The harness could just use the library's. Left alone
because that changes what the harness composes, and this commit only moves files.
