+++ Standard.Agents.Conformance/Brokers/ScriptedGeneratorBroker.cs
@@ -1,0 +1,25 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+using Standard.Agents.Brokers.Generators;
+
+namespace Standard.Agents.Conformance;
+
+public sealed class ScriptedGeneratorBroker : IGeneratorBroker
+{
+    private readonly IReadOnlyList<string> replies;
+    private int index;
+
+    public ScriptedGeneratorBroker(IReadOnlyList<string> replies) =>
+        this.replies = replies;
+
+    public ValueTask<string> GenerateAsync(string systemPrompt, string userPrompt)
+    {
+        string reply = this.replies[Math.Min(this.index, this.replies.Count - 1)];
+        this.index++;
+
+        return ValueTask.FromResult(reply);
+    }
+    }
