+++ Standard.Agents.Conformance/Models/Expectation.cs
@@ -1,0 +1,10 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+namespace Standard.Agents.Conformance;
+
+public sealed record Expectation(
+    string? Result,
+    string? ResultContains);
