+++ Standard.Agents/Brokers/Generators/GeneratorBroker.cs
@@ -6,6 +6,7 @@
 using System.Net.Http.Headers;
 using System.Text.Json;
 using RESTFulSense.Clients;
+using Standard.Agents.Models.Brokers.Generators;
 
 namespace Standard.Agents.Brokers.Generators;
 
@@ -81,21 +82,4 @@ public sealed class GeneratorBroker : IGeneratorBroker
                 ValueTask.FromResult(JsonSerializer.Serialize(value, jsonOptions)),
             deserializationFunction: json =>
                 ValueTask.FromResult(JsonSerializer.Deserialize<TResult>(json, jsonOptions)!));
-
-    private sealed record ChatCompletionRequest(
-    string Model,
-    IReadOnlyList<ChatMessage> Messages,
-    bool Stream,
-    double Temperature,
-    int MaxTokens);
-
-    private sealed record ChatMessage(
-        string Role,
-        string Content);
-
-    private sealed record ChatCompletionResponse(
-        IReadOnlyList<ChatChoice> Choices);
-
-    private sealed record ChatChoice(
-        ChatMessage Message);
     }
