+++ Standard.Agents/Brokers/Verifiers/VerifierBroker.cs
@@ -7,6 +7,7 @@ using System.Globalization;
 using System.Net.Http.Headers;
 using System.Text.Json;
 using RESTFulSense.Clients;
+using Standard.Agents.Models.Brokers.Verifiers;
 
 namespace Standard.Agents.Brokers.Verifiers;
 
@@ -87,21 +88,4 @@ CultureInfo.InvariantCulture);
                 ValueTask.FromResult(JsonSerializer.Serialize(value, jsonOptions)),
             deserializationFunction: json =>
                 ValueTask.FromResult(JsonSerializer.Deserialize<TResult>(json, jsonOptions)!));
-
-    private sealed record ChatCompletionRequest(
-        string Model,
-        IReadOnlyList<ChatMessage> Messages,
-        bool Stream,
-        double Temperature,
-        int MaxTokens);
-
-    private sealed record ChatMessage(
-        string Role,
-        string Content);
-
-    private sealed record ChatCompletionResponse(
-        IReadOnlyList<ChatChoice> Choices);
-
-    private sealed record ChatChoice(
-        ChatMessage Message);
     }
