+++ Standard.Agents/Models/Brokers/Classifiers/ChatCompletionRequest.cs
@@ -1,0 +1,13 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+namespace Standard.Agents.Models.Brokers.Classifiers;
+
+internal sealed record ChatCompletionRequest(
+    string Model,
+    IReadOnlyList<ChatMessage> Messages,
+    bool Stream,
+    double Temperature,
+    int MaxTokens);
