+++ Standard.Agents/Models/Brokers/Generators/ChatCompletionResponse.cs
@@ -1,0 +1,9 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+namespace Standard.Agents.Models.Brokers.Generators;
+
+internal sealed record ChatCompletionResponse(
+    IReadOnlyList<ChatChoice> Choices);
