+++ Standard.Agents/Models/Brokers/Generators/ChatMessage.cs
@@ -1,0 +1,10 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+namespace Standard.Agents.Models.Brokers.Generators;
+
+internal sealed record ChatMessage(
+    string Role,
+    string Content);
