+++ Standard.Agents/Models/Brokers/Mcps/JsonRpcRequest.cs
@@ -1,0 +1,14 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+using System.Text.Json.Serialization;
+
+namespace Standard.Agents.Models.Brokers.Mcps;
+
+internal sealed record JsonRpcRequest(
+    [property: JsonPropertyName("jsonrpc")] string JsonRpc,
+    int Id,
+    string Method,
+    ToolCallParams Params);
