+++ Standard.Agents/Models/Brokers/Mcps/ToolCallContent.cs
@@ -1,0 +1,10 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+namespace Standard.Agents.Models.Brokers.Mcps;
+
+internal sealed record ToolCallContent(
+    string Type,
+    string Text);
