+++ Standard.Agents/Models/Brokers/Verifiers/ChatChoice.cs
@@ -1,0 +1,9 @@
+// ---------------------------------------------------------------
+// Copyright (c) Hassan Habib All rights reserved.
+// Licensed under the The Standard Software License (TSSL)
+// ---------------------------------------------------------------
+
+namespace Standard.Agents.Models.Brokers.Verifiers;
+
+internal sealed record ChatChoice(
+    ChatMessage Message);
